requirements.txt
pyTelegramBotAPI==4.9.0
beautifulsoup4==4.11.1
requests==2.28.2
fake-useragent==1.1.3
russianLosses2==4.5
requests==2.28.2